[#179] refactor: 사용하지 않는 dto 삭제, 유즈케이스에 따라 메서드 위치 리팩터링
