+++ src/main/java/com/kiero/schedule/adapter/in/web/ScheduleController.java
@@ -72,7 +72,7 @@ public class ScheduleController {
 	public ResponseEntity<SuccessResponse<TodayScheduleResponse>> updateAndGetTodaySchedule(
 		@CurrentMember CurrentAuth currentAuth
 	) {
-		TodayScheduleResponse response = scheduleCommandUseCase.getTodaySchedule(currentAuth.memberId());
+		TodayScheduleResponse response = scheduleQueryUseCase.getTodaySchedule(currentAuth.memberId());
 		return ResponseEntity.ok()
 			.body(SuccessResponse.of(ScheduleSuccessCode.TODAY_SCHEDULE_GET_SUCCESS, response));
 	}
