--- src/main/java/com/kiero/schedule/application/dto/DiscardedScheduleDto.java
@@ -1,12 +1,0 @@
-package com.kiero.schedule.application.dto;
-
-import java.time.LocalDate;
-
-import com.kiero.schedule.domain.enums.DayOfWeek;
-
-public record DiscardedScheduleDto(
-	Long scheduleId,
-	LocalDate date,
-	DayOfWeek dayOfWeek
-) {
-}
