--- src/main/java/com/kiero/schedule/application/dto/RecurringScheduleDto.java
@@ -1,16 +1,0 @@
-package com.kiero.schedule.application.dto;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public record RecurringScheduleDto(
-	Long scheduleId,
-	LocalTime startTime,
-	LocalTime endTime,
-	String name,
-	String colorCode,
-	String dayOfWeek,
-	LocalDate repeatStartDate,
-	LocalDate repeatEndDate
-) {
-}
